+++ src/index.d.ts
@@ -1,17 +1,89 @@
-export interface AriesOptions {
+export interface SwaggerV2Definition {
+  type?: string;
+  required?: string[];
+  $ref?: string;
+  example?: string;
+  description?: string;
+  enum?: string[];
+  format?: string;
+  properties?: Record<string, SwaggerV2Definition>;
+  items?: Record<string, SwaggerV2Definition>;
+  [key: string]: any;
+}
+
+export interface SwaggerV2PathParameter {
+  name: string;
+  in: string;
+  description: string;
+  type?: string;
+  required: boolean;
+  format?: string;
+  schema?: SwaggerV2Definition;
+  [key: string]: any;
+}
+
+export interface SwaggerV2PathResponse {
+  description?: string;
+  schema?: SwaggerV2Definition;
+  headers?: Record<string, SwaggerV2Definition>;
+  [key: string]: any;
+}
+
+export interface SwaggerV2Path {
+  tags: string[];
+  description: string;
+  parameters: SwaggerV2PathParameter[];
+  responses: Record<string, SwaggerV2PathResponse>
+  [key: string]: any;
+}
+
+export interface SwaggerV2 {
+  swagger: "2.0";
+  basePath: string;
+  tags: {
+    name: string;
+    description?: string;
+    [key: string]: any;
+  }[];
+  paths: Record<string, Record<string, SwaggerV2Path>>;
+  definitions: Record<string, SwaggerV2Definition>;
+  [key: string]: any;
+}
+
+export interface RequiredOptions {
   url: string;
+}
+
+export interface ToTsOptions extends RequiredOptions {
   output: string;
-  autoMock: boolean;
-  resTemplate: string;
-  autoRequired: boolean;
-  formatProp: (prop: string) => string
+  autoRequired?: boolean;
+  formatProp?: (prop: string) => string
 }
 
-export interface PluginParams {
-  options: AriesOptions;
-  swagger: any
+export interface ToMdOptions extends RequiredOptions {
+  output: string;
+  autoMock?: boolean;
+  resTemplate?: string;
 }
 
-export declare function toTs(params: PluginParams): Promise<string>
-export declare function toMd(params: PluginParams): Promise<string>
-export declare function toMock(params: PluginParams): Promise<string>
+export interface ToMockOptions extends RequiredOptions {
+  output: string;
+  autoMock?: boolean;
+  resTemplate?: string;
+}
+
+export interface MockServerOptions extends RequiredOptions {
+  autoMock?: boolean;
+  resTemplate?: string;
+  port?: number
+}
+
+export interface PluginParams<T> {
+  options: T;
+  swagger: SwaggerV2
+}
+
+export declare function toTs(params: PluginParams<ToTsOptions>): Promise<string>
+export declare function toMd(params: PluginParams<ToMdOptions>): Promise<string>
+export declare function toMock(params: PluginParams<ToMockOptions>): Promise<string>
+export declare function mockServer(params: { options: MockServerOptions }): Promise<void>